More unit tests

+++ landmatrix/models/activity.py
@@ -9,7 +9,6 @@ from django.core.exceptions import ObjectDoesNotExist
 from django.utils import timezone
 from django.conf import settings
 
-from landmatrix.models.default_string_representation import DefaultStringRepresentation
 from landmatrix.models.activity_attribute_group import ActivityAttribute
 from landmatrix.models.investor import Investor, InvestorActivityInvolvement, InvestorBase, \
     HistoricalInvestorVentureInvolvement
@@ -30,12 +29,9 @@ class ActivityQuerySet(models.QuerySet):
             return self.filter(fk_status_id__in=ActivityBase.PUBLIC_STATUSES)
 
     def public_or_deleted(self, user=None):
-        statuses = ActivityBase.PUBLIC_STATUSES + (
-            ActivityBase.STATUS_DELETED,
-        )
+        statuses = ActivityBase.PUBLIC_STATUSES + (ActivityBase.STATUS_DELETED, )
         if user and user.is_authenticated:
-            return self.filter(models.Q(fk_status_id__in=statuses) |
-                        models.Q(history_user=user))
+            return self.filter(models.Q(fk_status_id__in=statuses) | models.Q(history_user=user))
         else:
             return self.filter(fk_status_id__in=statuses)
 
@@ -75,7 +71,7 @@ class ActivityQuerySet(models.QuerySet):
         return self.filter(fk_status_id=ActivityBase.STATUS_REJECTED)
 
 
-class ActivityBase(DefaultStringRepresentation, models.Model):
+class ActivityBase(models.Model):
     ACTIVITY_IDENTIFIER_DEFAULT = 2147483647  # Max safe int
 
     # FIXME: Replace fk_status with Choice Field
@@ -243,7 +239,6 @@ class ActivityBase(DefaultStringRepresentation, models.Model):
     def operational_stakeholder(self):
         if self.involvements.count() == 0:
             return
-            raise ObjectDoesNotExist('No OP for activity %s: %s' % (str(self), str(involvement)))
         else:
             involvement = self.involvements.order_by('-id')[0]
         return involvement.fk_investor
@@ -303,12 +298,9 @@ class ActivityBase(DefaultStringRepresentation, models.Model):
             country = country.first()
             try:
                 return Country.objects.defer('geom').get(id=country.value)
-            except Country.DoesNotExist:
+            except Country.DoesNotExist:  # pragma: no cover
                 return None
-            # Deprecated: Was necessary because of wrong values in the database
-            except:
-                return None
-        else:
+        else:  # pragma: no cover
             return None
 
     @property
@@ -364,7 +356,7 @@ class ActivityBase(DefaultStringRepresentation, models.Model):
                 country = Country.objects.defer('geom').select_related('fk_region')
                 country = country.get(id=country_id)
                 countries.append(country)
-            except Country.DoesNotExist:
+            except Country.DoesNotExist:  # pragma: no cover
                 pass
         return countries
 
@@ -475,20 +467,12 @@ class ActivityBase(DefaultStringRepresentation, models.Model):
         # 4B. Invalid Parent companies/investors?
         if self.has_invalid_operating_company(involvements) and self.has_invalid_parents(involvements):
             return False
-        # 6. High income country?
+        # 5. High income country?
         if self.is_high_income_target_country():
             return False
         return True
 
     def get_not_public_reason(self):
-        # Presets:
-        # >= 2000
-        # Size given and size > 200 ha
-        # 5. has subinvestors
-        # 6. has valid investor (not unknown)
-        # 7. Intention is not Mining
-        # 8. Target country is no high income country
-
         # 1. Flag „not public“ set?
         if self.has_flag_not_public():
             return '1. Flag not public set'
@@ -499,16 +483,13 @@ class ActivityBase(DefaultStringRepresentation, models.Model):
         involvements = self.involvements.all()
         if involvements.count() == 0:
             return '3. involvements missing'
-        # 4. Invalid Operating company name?
-        if self.has_invalid_operating_company(involvements):
-            return ''
-        # 4. Invalid Operating company name?
-        # 5. Invalid Parent companies/investors?
+        # 4A. Invalid Operating company name?
+        # 4B. Invalid Parent companies/investors?
         if self.has_invalid_operating_company(involvements) and self.has_invalid_parents(involvements):
-            return '4. Invalid Operating company name or 5. Invalid Parent companies/investors'
-        # 6. High income country
+            return '4. Invalid Operating company name and Invalid Parent companies/investors'
+        # 5. High income country?
         if self.is_high_income_target_country():
-            return '6. High income country'
+            return '5. High income country'
         return 'Filters passed (public)'
 
     def is_high_income_target_country(self):
@@ -527,7 +508,7 @@ class ActivityBase(DefaultStringRepresentation, models.Model):
 
     def has_invalid_operating_company(self, involvements):
         for i in involvements:
-            if not i.fk_investor:
+            if not i.fk_investor:  # pragma: no cover
                 continue
             investor_name = i.fk_investor.name
             invalid_name = "^(unknown|unnamed)( \(([, ]*(unnamed investor [0-9]+)*)+\))?$"
@@ -537,7 +518,7 @@ class ActivityBase(DefaultStringRepresentation, models.Model):
 
     def has_invalid_parents(self, involvements):
         for i in involvements:
-            if not i.fk_investor:
+            if not i.fk_investor:  # pragma: no cover
                 continue
             # Operating company name given?
             # investor_name = i.fk_investor.name
@@ -598,7 +579,7 @@ class ActivityBase(DefaultStringRepresentation, models.Model):
                 init_dates.append(implementation_stati[0].date)
         if init_dates:
             return min(init_dates)
-        else:
+        else:  # pragma: no cover
             return None
 
     def get_deal_scope(self):
@@ -634,7 +615,7 @@ class ActivityBase(DefaultStringRepresentation, models.Model):
         elif len(target_countries) > 0 and len(investor_countries) == 0:
             # treat deals without investor country as transnational
             return "transnational"
-        else:
+        else:  # pragma: no cover
             return None
 
     def format_investors(self, investors):
@@ -663,14 +644,14 @@ class ActivityBase(DefaultStringRepresentation, models.Model):
         try:
             activity = HistoricalActivity.objects.filter(activity_identifier=self.activity_identifier).latest()
             return activity.history_date
-        except:
+        except:  # pragma: no cover
             return None
 
     def get_updated_user(self):
         try:
             activity = HistoricalActivity.objects.filter(activity_identifier=self.activity_identifier).latest()
             return activity.history_user.id
-        except:
+        except:  # pragma: no cover
             return None
 
     def get_fully_updated_date(self):
@@ -678,7 +659,7 @@ class ActivityBase(DefaultStringRepresentation, models.Model):
             activity = HistoricalActivity.objects.filter(activity_identifier=self.activity_identifier,
                                                          fully_updated=True).latest()
             return activity.history_date
-        except:
+        except:  # pragma: no cover
             return None
 
     def get_fully_updated_user(self):
@@ -686,7 +667,7 @@ class ActivityBase(DefaultStringRepresentation, models.Model):
             activity = HistoricalActivity.objects.filter(activity_identifier=self.activity_identifier,
                                                          fully_updated=True).latest()
             return activity.history_user.id
-        except:
+        except:  # pragma: no cover
             return None
 
 
@@ -816,7 +797,8 @@ class HistoricalActivity(ActivityBase):
     objects = HistoricalActivityQuerySet.as_manager()
 
     def approve_change(self, user=None, comment=None):
-        assert self.fk_status_id == HistoricalActivity.STATUS_PENDING
+        if self.fk_status_id != HistoricalActivity.STATUS_PENDING:  # pragma: no cover
+            return
 
         # Only approvals of administrators should go public
         if user.has_perm('landmatrix.change_activity'):
@@ -844,7 +826,8 @@ class HistoricalActivity(ActivityBase):
             self.changesets.create(fk_user=user, comment=comment)
 
     def reject_change(self, user=None, comment=None):
-        assert self.fk_status_id == HistoricalActivity.STATUS_PENDING
+        if self.fk_status_id != HistoricalActivity.STATUS_PENDING:  # pragma: no cover
+            return
 
         # Only rejections of administrators should go public
         if user.has_perm('landmatrix.change_activity'):
@@ -862,7 +845,8 @@ class HistoricalActivity(ActivityBase):
             self.changesets.create(fk_user=user, comment=comment)
 
     def approve_delete(self, user=None, comment=None):
-        assert self.fk_status_id == HistoricalActivity.STATUS_TO_DELETE
+        if self.fk_status_id != HistoricalActivity.STATUS_TO_DELETE:  # pragma: no cover
+            return
 
         # Only approvals of administrators should be deleted
         if user.has_perm('landmatrix.delete_activity'):
@@ -873,7 +857,8 @@ class HistoricalActivity(ActivityBase):
             self.changesets.create(fk_user=user, comment=comment)
 
     def reject_delete(self, user=None, comment=None):
-        assert self.fk_status_id == HistoricalActivity.STATUS_TO_DELETE
+        if self.fk_status_id != HistoricalActivity.STATUS_TO_DELETE:  # pragma: no cover
+            return
 
         # Only approvals of administrators should be deleted
         if user.has_perm('landmatrix.delete_activity'):
@@ -882,7 +867,7 @@ class HistoricalActivity(ActivityBase):
 
             try:
                 investor = self.involvements.all()[0].fk_investor
-            except IndexError:
+            except IndexError:  # pragma: no cover
                 pass
             else:
                 investor.reject_change(user=user, comment=comment)
@@ -947,9 +932,9 @@ class HistoricalActivity(ActivityBase):
         elif self.fk_status_id == self.STATUS_REJECTED:
             # Activity add has been rejected?
             activities = HistoricalActivity.objects.filter(activity_identifier=self.activity_identifier)
-            if len(activities) == 1:
+            if activity and len(activities) == 1:
                 activity.delete()
-                return
+            return
 
         if not activity:
             activity = Activity.objects.create(
